+++ src/graphql/network.rs
@@ -700,7 +700,7 @@ impl NetworkQuery {
             |after, before, first, last| async move {
                 let (conn_iter, cursor, size) = get_filtered_iter(
                     &db.conn_store()?,
-                    &key_prefix,
+                    Some(&key_prefix),
                     &filter,
                     &after,
                     &before,
@@ -718,7 +718,7 @@ impl NetworkQuery {
 
                 let (dns_iter, cursor, _) = get_filtered_iter(
                     &db.dns_store()?,
-                    &key_prefix,
+                    Some(&key_prefix),
                     &filter,
                     &after,
                     &before,
@@ -736,7 +736,7 @@ impl NetworkQuery {
 
                 let (http_iter, cursor, _) = get_filtered_iter(
                     &db.http_store()?,
-                    &key_prefix,
+                    Some(&key_prefix),
                     &filter,
                     &after,
                     &before,
@@ -754,7 +754,7 @@ impl NetworkQuery {
 
                 let (rdp_iter, cursor, _) = get_filtered_iter(
                     &db.rdp_store()?,
-                    &key_prefix,
+                    Some(&key_prefix),
                     &filter,
                     &after,
                     &before,
@@ -772,7 +772,7 @@ impl NetworkQuery {
 
                 let (ntlm_iter, cursor, _) = get_filtered_iter(
                     &db.ntlm_store()?,
-                    &key_prefix,
+                    Some(&key_prefix),
                     &filter,
                     &after,
                     &before,
@@ -790,7 +790,7 @@ impl NetworkQuery {
 
                 let (kerberos_iter, cursor, _) = get_filtered_iter(
                     &db.kerberos_store()?,
-                    &key_prefix,
+                    Some(&key_prefix),
                     &filter,
                     &after,
                     &before,
@@ -808,7 +808,7 @@ impl NetworkQuery {
 
                 let (ssh_iter, cursor, _) = get_filtered_iter(
                     &db.ssh_store()?,
-                    &key_prefix,
+                    Some(&key_prefix),
                     &filter,
                     &after,
                     &before,
@@ -826,7 +826,7 @@ impl NetworkQuery {
 
                 let (dce_rpc_iter, cursor, _) = get_filtered_iter(
                     &db.dce_rpc_store()?,
-                    &key_prefix,
+                    Some(&key_prefix),
                     &filter,
                     &after,
                     &before,